First attempt at a MotorGroup.
For now, motor groups will extend speed controller
and implement subsystem. I am doing this because
extending motor seems not helpful, it is better
to construct a set of motors, then use that as
a single speed controller.
Basically, make a motor group, then use it as the
speed controller for a motor, an accel motor, or
an encoded motor (this will be very important for
keeping motors in sync).